+++ pyro/poutine/subsample_messenger.py
@@ -1,0 +1,121 @@
+from __future__ import absolute_import, division, print_function
+
+import torch
+
+from pyro.distributions.distribution import Distribution
+
+from .indep_messenger import IndepMessenger
+from .runtime import apply_stack, am_i_wrapped
+
+
+class _Subsample(Distribution):
+    """
+    Randomly select a subsample of a range of indices.
+
+    Internal use only. This should only be used by `iarange`.
+    """
+
+    def __init__(self, size, subsample_size, use_cuda=None):
+        """
+        :param int size: the size of the range to subsample from
+        :param int subsample_size: the size of the returned subsample
+        :param bool use_cuda: whether to use cuda tensors
+        """
+        self.size = size
+        self.subsample_size = subsample_size
+        self.use_cuda = torch.Tensor().is_cuda if use_cuda is None else use_cuda
+
+    def sample(self, sample_shape=torch.Size()):
+        """
+        :returns: a random subsample of `range(size)`
+        :rtype: torch.LongTensor
+        """
+        if sample_shape:
+            raise NotImplementedError
+        subsample_size = self.subsample_size
+        if subsample_size is None or subsample_size > self.size:
+            subsample_size = self.size
+        if subsample_size == self.size:
+            result = torch.LongTensor(list(range(self.size)))
+        else:
+            # torch.randperm does not have a CUDA implementation
+            result = torch.randperm(self.size, device=torch.device('cpu'))[:self.subsample_size]
+        return result.cuda() if self.use_cuda else result
+
+    def log_prob(self, x):
+        # This is zero so that iarange can provide an unbiased estimate of
+        # the non-subsampled log_prob.
+        result = torch.zeros(1)
+        return result.cuda() if self.use_cuda else result
+
+
+class SubsampleMessenger(IndepMessenger):
+    """
+    Drop-in replacement for irange and iarange, including subsampling!
+    """
+
+    def __init__(self, name, size=None, subsample_size=None, subsample=None, dim=None, use_cuda=None, sites=None):
+        super(SubsampleMessenger, self).__init__(name, size, dim, sites)
+        self._size = size
+        self.subsample_size = subsample_size
+        self.indices = subsample
+        self.use_cuda = use_cuda
+
+        if am_i_wrapped():
+            self._size, self.subsample_size, self.indices = self._subsample(
+                self.name, self._size, self.subsample_size,
+                self.indices, self.use_cuda)
+            self.size = self.subsample_size
+
+    def _subsample(self, name, size=None, subsample_size=None, subsample=None, use_cuda=None):
+        """
+        Helper function for iarange and irange. See their docstrings for details.
+        """
+        if size is None:
+            assert subsample_size is None
+            assert subsample is None
+            size = -1  # This is PyTorch convention for "arbitrary size"
+            subsample_size = -1
+        elif subsample is None:
+            msg = {
+                "type": "sample",
+                "name": name,
+                "fn": _Subsample(size, subsample_size, use_cuda),
+                "is_observed": False,
+                "args": (),
+                "kwargs": {},
+                "value": None,
+                "infer": {},
+                "scale": 1.0,
+                "cond_indep_stack": (),
+                "done": False,
+                "stop": False,
+                "continuation": None
+            }
+            apply_stack(msg)
+            subsample = msg["value"]
+
+        if subsample_size is None:
+            subsample_size = len(subsample)
+        elif subsample is not None and subsample_size != len(subsample):
+            raise ValueError("subsample_size does not match len(subsample), {} vs {}.".format(
+                subsample_size, len(subsample)) +
+                " Did you accidentally use different subsample_size in the model and guide?")
+
+        return size, subsample_size, subsample
+
+    def __enter__(self):
+        if self.indices is None and self.size is None:
+            self._size, self.subsample_size, self.indices = self._subsample(
+                self.name, self._size, self.subsample_size, self.indices, self.use_cuda)
+        self.size = self.subsample_size
+        return super(SubsampleMessenger, self).__enter__()
+
+    def _reset(self):
+        self.subsample = None
+        super(SubsampleMessenger, self)._reset()
+
+    def _process_message(self, msg):
+        super(SubsampleMessenger, self)._process_message(msg)
+        if self._installed:
+            msg["scale"] = (self._size / self.subsample_size) * msg["scale"]
